[testsuite][gha] use poetry everywhere and refactor testsuite
