+++ .github/actions/run-faucet-tests/action.yaml
@@ -28,7 +28,7 @@ runs:
       uses: aptos-labs/aptos-core/.github/actions/rust-setup@main
 
     # Install Poetry.
-    - uses: snok/install-poetry@v1
+    - uses: snok/install-poetry@d45b6d76012debf457ab49dffc7fb7b2efe8071d # pin@v1.3.3
       with:
         version: 1.2.2
