+++ .github/workflows/cli-e2e-tests.yaml
@@ -36,7 +36,7 @@ jobs:
           GIT_CREDENTIALS: ${{ secrets.GIT_CREDENTIALS }}
 
       # Install Poetry.
-      - uses: snok/install-poetry@v1
+      - uses: snok/install-poetry@d45b6d76012debf457ab49dffc7fb7b2efe8071d # pin@v1.3.3
         with:
           version: 1.2.2
