+++ .github/workflows/forge-stable.yaml
@@ -77,13 +77,19 @@ jobs:
 
       - uses: actions/setup-python@13ae5bb136fac2878aff31522b9efb785519f984 # pin@v4
 
-      - name: Install python deps
-        run: pip3 install click==8.1.3 psutil==5.9.1
+      - uses: snok/install-poetry@d45b6d76012debf457ab49dffc7fb7b2efe8071d # pin@v1.3.3
+        with:
+          version: 1.2.2
+      - run: poetry install
+        shell: bash
+        working-directory: testsuite
 
       - name: Determine image tag
         id: determine-test-image-tag
         # forge relies on the default and failpoints variants
-        run: python3 testsuite/find_latest_image.py --variant failpoints --variant performance
+        run: ./testrun find_latest_image.py --variant failpoints --variant performance
+        shell: bash
+        working-directory: testsuite
 
       - name: Write summary
         run: |
@@ -283,7 +289,7 @@ jobs:
       FORGE_TEST_SUITE: state_sync_perf_fullnodes_apply_outputs
       POST_TO_SLACK: ${{ needs.determine-test-metadata.outputs.BRANCH == 'main' }} # only post to slack on main branch
 
-### Additional three-region tests. Eventually all consensus-related tests should migrate to three-regions.
+  ### Additional three-region tests. Eventually all consensus-related tests should migrate to three-regions.
 
   run-forge-land-blocking-three-region:
     if: ${{ github.event_name != 'pull_request' }}
