--- testsuite/requirements.txt
@@ -1,2 +1,0 @@
-psutil==5.9.1
-click==8.1.3
