+++ testsuite/test_framework/README.md
@@ -1,0 +1,5 @@
+# Test Framework
+
+Library for common python test framework utilities.
+
+As opposed to to unit tests of core Rust components, which should be written in Rust along with the code, these python test utilities are for integration and e2e testing.
