+++ testsuite/test_framework/process.py
@@ -1,3 +1,5 @@
+# A wrapper around operating system processes
+
 from __future__ import annotations
 
 import atexit
@@ -57,3 +59,41 @@ class SystemProcesses(Processes):
 
     def user(self) -> str:
         return pwd.getpwuid(os.getuid())[0]
+
+
+@dataclass
+class FakeProcess(Process):
+    _name: str
+    _ppid: int
+
+    def name(self) -> str:
+        return self._name
+
+    def ppid(self) -> int:
+        return self._ppid
+
+
+class FakeProcesses(Processes):
+    def __init__(self) -> None:
+        self.exit_callbacks = []
+
+    def processes(self) -> Generator[Process, None, None]:
+        yield FakeProcess("concensus", 1)
+
+    def get_pid(self) -> int:
+        return 2
+
+    def spawn(self, target: Callable[[], None]) -> Process:
+        return FakeProcess("child", 2)
+
+    def atexit(self, callback: Callable[[], None]) -> None:
+        return self.exit_callbacks.append(callback)
+
+    def user(self) -> str:
+        return "perry"
+
+
+class SpyProcesses(FakeProcesses):
+    def run_atexit(self) -> None:
+        for callback in self.exit_callbacks:
+            callback()
